more promises and async/wait
simplifies api a bit.

fade() resolves when it is done fading, and it no longer takes a 3rd "done callback" param.
there is now a delay(ms), resolving when done waiting.
confirm() returns a promise resulting in whether it was confirmed or not.
there is a new reveal() function that just shows content. instead of going through a Loader. simplifies the code.